:building_construction: move graphQl folder
