+++ frontend/__tests__/components/users/signin/SignIn.test.tsx
@@ -15,7 +15,7 @@ import {
 import { MockedProvider } from "@apollo/client/testing";
 import { toast } from "react-hot-toast";
 import SignIn from "@/components/users/signin/SignIn";
-import { mutationUserLogin, queryMeContext } from "@/components/graphql/Users";
+import { mutationUserLogin, queryMeContext } from "@/graphql/Users";
 
 // Mocks React-Hot-Toast
 jest.mock("react-hot-toast");
