+++ frontend/src/config/userAuth.ts
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { useRouter } from "next/router";
 import { useQuery } from "@apollo/client";
-import { queryMeContext } from "@/components/graphql/Users";
+import { queryMeContext } from "@/graphql/Users";
 import { UserContextTypes } from "@/types/UserTypes";
 
 export function useAuth(privatePages: string[]) {
