+++ frontend/src/graphql/Ads.ts
