+++ frontend/src/graphql/Categories.ts
